changed project signing key changed stock_quantity to decimal

+++ WooCommerce/Legacy/Product.cs
@@ -161,8 +161,8 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// <summary>
         /// Stock quantity. If is a variable product this value will be used to control stock for all variations, unless you define stock at variation level.
         /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-        public int? stock_quantity { get; set; }
+        [DataMember(EmitDefaultValue = false)]        
+        public decimal? stock_quantity { get; set; }
 
         /// <summary>
         /// Controls whether or not the product is listed as “in stock” or “out of stock” on the frontend.
@@ -776,7 +776,7 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// Stock quantity. If is a variable variation this value will be used to control stock for all variations, unless you define stock at variation level.
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? stock_quantity { get; set; }
+        public decimal? stock_quantity { get; set; }
 
         /// <summary>
         /// Controls whether or not the variation is listed as “in stock” or “out of stock” on the frontend.
